Minor cleanups.
